src/daily_news/requirements.txt
feedparser
scikit-learn
mcp-server
fastapi
python-dotenv
